+++ src/api/api.tsx
@@ -29,116 +29,146 @@ import {
   User,
   UserCreate,
   Warning,
+  Badge,
+  GroupList,
 } from 'types/Types';
+
+export const AUTH_ENDPOINT = 'auth';
+export const BADGES_ENDPOINT = 'badge';
+export const CATEGORIES_ENDPOINT = 'category';
+export const CHEATSHEETS_ENDPOINT = 'cheatsheet';
+export const EVENTS_ENDPOINT = 'events';
+export const EVENT_REGISTRATIONS_ENDPOINT = 'users';
+export const FORMS_ENDPOINT = 'forms';
+export const GROUPS_ENDPOINT = 'group';
+export const JOBPOSTS_ENDPOINT = 'jobpost';
+export const ME_ENDPOINT = 'me';
+export const MEMBERSHIPS_ENDPOINT = 'membership';
+export const NEWS_ENDPOINT = 'news';
+export const NOTIFICATIONS_ENDPOINT = 'notification';
+export const PAGES_ENDPOINT = 'page';
+export const SHORT_LINKS_ENDPOINT = 'short-link';
+export const USERS_ENDPOINT = 'user';
+export const WARNINGS_ENDPOINT = 'warning';
 
 export default {
   // Auth
-  createUser: (item: UserCreate) => IFetch<RequestResponse>({ method: 'POST', url: 'user/', data: item, withAuth: false }),
+  createUser: (item: UserCreate) => IFetch<RequestResponse>({ method: 'POST', url: `${USERS_ENDPOINT}/`, data: item, withAuth: false }),
   authenticate: (username: string, password: string) =>
     IFetch<LoginRequestResponse>({
       method: 'POST',
-      url: 'auth/login/',
+      url: `${AUTH_ENDPOINT}/login/`,
       data: { user_id: username, password: password },
       withAuth: false,
     }),
   forgotPassword: (email: string) =>
-    IFetch<RequestResponse>({ method: 'POST', url: 'auth/rest-auth/password/reset/', data: { email: email }, withAuth: false }),
+    IFetch<RequestResponse>({ method: 'POST', url: `${AUTH_ENDPOINT}/rest-auth/password/reset/`, data: { email: email }, withAuth: false }),
 
   // Events
-  getEvent: (eventId: number) => IFetch<Event>({ method: 'GET', url: `events/${String(eventId)}/` }),
-  getEvents: (filters?: any) => IFetch<PaginationResponse<EventCompact>>({ method: 'GET', url: `events/`, data: filters || {} }),
-  createEvent: (item: EventRequired) => IFetch<Event>({ method: 'POST', url: `events/`, data: item }),
-  updateEvent: (eventId: number, item: Partial<Event>) => IFetch<Event>({ method: 'PUT', url: `events/${String(eventId)}/`, data: item }),
-  deleteEvent: (eventId: number) => IFetch<RequestResponse>({ method: 'DELETE', url: `events/${String(eventId)}/` }),
+  getEvent: (eventId: number) => IFetch<Event>({ method: 'GET', url: `${EVENTS_ENDPOINT}/${String(eventId)}/` }),
+  getEvents: (filters?: any) => IFetch<PaginationResponse<EventCompact>>({ method: 'GET', url: `${EVENTS_ENDPOINT}/`, data: filters || {} }),
+  createEvent: (item: EventRequired) => IFetch<Event>({ method: 'POST', url: `${EVENTS_ENDPOINT}/`, data: item }),
+  updateEvent: (eventId: number, item: Partial<Event>) => IFetch<Event>({ method: 'PUT', url: `${EVENTS_ENDPOINT}/${String(eventId)}/`, data: item }),
+  deleteEvent: (eventId: number) => IFetch<RequestResponse>({ method: 'DELETE', url: `${EVENTS_ENDPOINT}/${String(eventId)}/` }),
   putAttended: (eventId: number, item: { has_attended: boolean }, userId: string) =>
-    IFetch<RequestResponse>({ method: 'PUT', url: `events/${String(eventId)}/users/${userId}/`, data: item }),
-  getRegistration: (eventId: number, userId: string) => IFetch<Registration>({ method: 'GET', url: `events/${String(eventId)}/users/${userId}/` }),
-  getEventRegistrations: (eventId: number) => IFetch<Array<Registration>>({ method: 'GET', url: `events/${String(eventId)}/users/` }),
+    IFetch<RequestResponse>({ method: 'PUT', url: `${EVENTS_ENDPOINT}/${String(eventId)}/${EVENT_REGISTRATIONS_ENDPOINT}/${userId}/`, data: item }),
+  getRegistration: (eventId: number, userId: string) =>
+    IFetch<Registration>({ method: 'GET', url: `${EVENTS_ENDPOINT}/${String(eventId)}/${EVENT_REGISTRATIONS_ENDPOINT}/${userId}/` }),
+  getEventRegistrations: (eventId: number) =>
+    IFetch<Array<Registration>>({ method: 'GET', url: `${EVENTS_ENDPOINT}/${String(eventId)}/${EVENT_REGISTRATIONS_ENDPOINT}/` }),
   createRegistration: (eventId: number, item: Partial<Registration>) =>
-    IFetch<Registration>({ method: 'POST', url: `events/${String(eventId)}/users/`, data: item }),
+    IFetch<Registration>({ method: 'POST', url: `${EVENTS_ENDPOINT}/${String(eventId)}/${EVENT_REGISTRATIONS_ENDPOINT}/`, data: item }),
   updateRegistration: (eventId: number, item: Partial<Registration>, userId: string) =>
-    IFetch<Registration>({ method: 'PUT', url: `events/${String(eventId)}/users/${userId}/`, data: item }),
-  deleteRegistration: (eventId: number, userId: string) => IFetch<RequestResponse>({ method: 'DELETE', url: `events/${String(eventId)}/users/${userId}/` }),
+    IFetch<Registration>({ method: 'PUT', url: `${EVENTS_ENDPOINT}/${String(eventId)}/${EVENT_REGISTRATIONS_ENDPOINT}/${userId}/`, data: item }),
+  deleteRegistration: (eventId: number, userId: string) =>
+    IFetch<RequestResponse>({ method: 'DELETE', url: `${EVENTS_ENDPOINT}/${String(eventId)}/${EVENT_REGISTRATIONS_ENDPOINT}/${userId}/` }),
 
   // Forms
-  getForm: (formId: string) => IFetch<Form>({ method: 'GET', url: `forms/${formId}/` }),
-  createForm: (item: Form) => IFetch<Form>({ method: 'POST', url: `forms/`, data: item }),
-  updateForm: (formId: string, item: Form) => IFetch<Form>({ method: 'PUT', url: `forms/${formId}/`, data: item }),
-  deleteForm: (formId: string) => IFetch<RequestResponse>({ method: 'DELETE', url: `forms/${formId}/` }),
+  getForm: (formId: string) => IFetch<Form>({ method: 'GET', url: `${FORMS_ENDPOINT}/${formId}/` }),
+  createForm: (item: Form) => IFetch<Form>({ method: 'POST', url: `${FORMS_ENDPOINT}/`, data: item }),
+  updateForm: (formId: string, item: Form) => IFetch<Form>({ method: 'PUT', url: `${FORMS_ENDPOINT}/${formId}/`, data: item }),
+  deleteForm: (formId: string) => IFetch<RequestResponse>({ method: 'DELETE', url: `${FORMS_ENDPOINT}/${formId}/` }),
 
   // Job posts
-  getJobPosts: (filters: any = {}) => IFetch<PaginationResponse<JobPost>>({ method: 'GET', url: `jobpost/`, data: filters }),
-  getJobPost: (id: number) => IFetch<JobPost>({ method: 'GET', url: `jobpost/${String(id)}/` }),
-  createJobPost: (item: JobPostRequired) => IFetch<JobPost>({ method: 'POST', url: `jobpost/`, data: item }),
-  putJobPost: (id: number, item: JobPostRequired) => IFetch<JobPost>({ method: 'PUT', url: `jobpost/${String(id)}/`, data: item }),
-  deleteJobPost: (id: number) => IFetch<RequestResponse>({ method: 'DELETE', url: `jobpost/${String(id)}/` }),
+  getJobPosts: (filters: any = {}) => IFetch<PaginationResponse<JobPost>>({ method: 'GET', url: `${JOBPOSTS_ENDPOINT}/`, data: filters }),
+  getJobPost: (id: number) => IFetch<JobPost>({ method: 'GET', url: `${JOBPOSTS_ENDPOINT}/${String(id)}/` }),
+  createJobPost: (item: JobPostRequired) => IFetch<JobPost>({ method: 'POST', url: `${JOBPOSTS_ENDPOINT}/`, data: item }),
+  putJobPost: (id: number, item: JobPostRequired) => IFetch<JobPost>({ method: 'PUT', url: `${JOBPOSTS_ENDPOINT}/${String(id)}/`, data: item }),
+  deleteJobPost: (id: number) => IFetch<RequestResponse>({ method: 'DELETE', url: `${JOBPOSTS_ENDPOINT}/${String(id)}/` }),
 
   // News
-  getNewsItem: (id: number) => IFetch<News>({ method: 'GET', url: `news/${String(id)}/` }),
-  getNewsItems: (filters?: any) => IFetch<PaginationResponse<News>>({ method: 'GET', url: `news/`, data: filters || {} }),
-  createNewsItem: (item: NewsRequired) => IFetch<News>({ method: 'POST', url: `news/`, data: item }),
-  putNewsItem: (id: number, item: NewsRequired) => IFetch<News>({ method: 'PUT', url: `news/${String(id)}/`, data: item }),
-  deleteNewsItem: (id: number) => IFetch<RequestResponse>({ method: 'DELETE', url: `news/${String(id)}/` }),
+  getNewsItem: (id: number) => IFetch<News>({ method: 'GET', url: `${NEWS_ENDPOINT}/${String(id)}/` }),
+  getNewsItems: (filters?: any) => IFetch<PaginationResponse<News>>({ method: 'GET', url: `${NEWS_ENDPOINT}/`, data: filters || {} }),
+  createNewsItem: (item: NewsRequired) => IFetch<News>({ method: 'POST', url: `${NEWS_ENDPOINT}/`, data: item }),
+  putNewsItem: (id: number, item: NewsRequired) => IFetch<News>({ method: 'PUT', url: `${NEWS_ENDPOINT}/${String(id)}/`, data: item }),
+  deleteNewsItem: (id: number) => IFetch<RequestResponse>({ method: 'DELETE', url: `${NEWS_ENDPOINT}/${String(id)}/` }),
 
   // User
-  getUserData: () => IFetch<User>({ method: 'GET', url: `user/userdata/` }),
-  getUsers: (filters?: any) => IFetch<PaginationResponse<User>>({ method: 'GET', url: `user/`, data: filters || {} }),
-  updateUserData: (userName: string, item: Partial<User>) => IFetch<User>({ method: 'PUT', url: `user/${userName}/`, data: item }),
+  getUserData: () => IFetch<User>({ method: 'GET', url: `${USERS_ENDPOINT}/${ME_ENDPOINT}/` }),
+  getUserBadges: (filters?: any) =>
+    IFetch<PaginationResponse<Badge>>({ method: 'GET', url: `${USERS_ENDPOINT}/${ME_ENDPOINT}/${BADGES_ENDPOINT}/`, data: filters || {} }),
+  getUserEvents: (filters?: any) =>
+    IFetch<PaginationResponse<EventCompact>>({ method: 'GET', url: `${USERS_ENDPOINT}/${ME_ENDPOINT}/${EVENTS_ENDPOINT}/`, data: filters || {} }),
+  getUserGroups: () => IFetch<Array<GroupList>>({ method: 'GET', url: `${USERS_ENDPOINT}/${ME_ENDPOINT}/${GROUPS_ENDPOINT}/` }),
+  getUsers: (filters?: any) => IFetch<PaginationResponse<User>>({ method: 'GET', url: `${USERS_ENDPOINT}/`, data: filters || {} }),
+  updateUserData: (userName: string, item: Partial<User>) => IFetch<User>({ method: 'PUT', url: `${USERS_ENDPOINT}/${userName}/`, data: item }),
   activateUser: (userName: string) => IFetch<RequestResponse>({ method: 'POST', url: `activate-user/`, data: { user_id: userName } }),
 
   // Notifications
-  getNotifications: (filters?: any) => IFetch<PaginationResponse<Notification>>({ method: 'GET', url: `notification/`, data: filters || {} }),
-  updateNotification: (id: number, item: { read: boolean }) => IFetch<Notification>({ method: 'PUT', url: `notification/${String(id)}/`, data: item }),
+  getNotifications: (filters?: any) => IFetch<PaginationResponse<Notification>>({ method: 'GET', url: `${NOTIFICATIONS_ENDPOINT}/`, data: filters || {} }),
+  updateNotification: (id: number, item: { read: boolean }) =>
+    IFetch<Notification>({ method: 'PUT', url: `${NOTIFICATIONS_ENDPOINT}/${String(id)}/`, data: item }),
 
   // Short links
-  getShortLinks: (filters?: any) => IFetch<Array<ShortLink>>({ method: 'GET', url: `short-link/`, data: filters || {} }),
-  createShortLink: (item: ShortLink) => IFetch<ShortLink>({ method: 'POST', url: `short-link/`, data: item }),
-  deleteShortLink: (slug: string) => IFetch<RequestResponse>({ method: 'DELETE', url: `short-link/${slug}/` }),
+  getShortLinks: (filters?: any) => IFetch<Array<ShortLink>>({ method: 'GET', url: `${SHORT_LINKS_ENDPOINT}/`, data: filters || {} }),
+  createShortLink: (item: ShortLink) => IFetch<ShortLink>({ method: 'POST', url: `${SHORT_LINKS_ENDPOINT}/`, data: item }),
+  deleteShortLink: (slug: string) => IFetch<RequestResponse>({ method: 'DELETE', url: `${SHORT_LINKS_ENDPOINT}/${slug}/` }),
 
   // Cheatsheet
   getCheatsheets: (study: Study, grade: number, filters?: any) => {
     const tempStudy = study === Study.DIGSEC ? 'DIGINC' : study;
     return IFetch<PaginationResponse<Cheatsheet>>({
       method: 'GET',
-      url: `cheatsheet/${tempStudy.toUpperCase()}/${String(grade)}/files/`,
+      url: `${CHEATSHEETS_ENDPOINT}/${tempStudy.toUpperCase()}/${String(grade)}/files/`,
       data: filters || {},
       withAuth: true,
     });
   },
 
   // Warning
-  getWarning: () => IFetch<Array<Warning>>({ method: 'GET', url: `warning/` }),
+  getWarning: () => IFetch<Array<Warning>>({ method: 'GET', url: `${WARNINGS_ENDPOINT}/` }),
 
   // Categories
-  getCategories: () => IFetch<Array<Category>>({ method: 'GET', url: `category/` }),
+  getCategories: () => IFetch<Array<Category>>({ method: 'GET', url: `${CATEGORIES_ENDPOINT}/` }),
 
   // Company form
   emailForm: (data: CompaniesEmail) => IFetch<RequestResponse>({ method: 'POST', url: `accept-form/`, data, withAuth: false }),
 
   // Badges
-  createUserBadge: (data: { badge_id: string }) => IFetch<RequestResponse>({ method: 'POST', url: `badge/`, data }),
+  createUserBadge: (data: { badge_id: string }) => IFetch<RequestResponse>({ method: 'POST', url: `${BADGES_ENDPOINT}/`, data }),
 
   //Membership
   getMemberships: (slug: string, filters?: any) =>
-    IFetch<PaginationResponse<Membership>>({ method: 'GET', url: `group/${slug}/membership/`, data: filters || {} }),
+    IFetch<PaginationResponse<Membership>>({ method: 'GET', url: `${GROUPS_ENDPOINT}/${slug}/${MEMBERSHIPS_ENDPOINT}/`, data: filters || {} }),
   createMembership: (slug: string, userId: string) =>
-    IFetch<Membership>({ method: 'POST', url: `group/${slug}/membership/`, data: { user: { user_id: userId } } }),
-  deleteMembership: (slug: string, userId: string) => IFetch<RequestResponse>({ method: 'DELETE', url: `group/${slug}/membership/${userId}/` }),
+    IFetch<Membership>({ method: 'POST', url: `${GROUPS_ENDPOINT}/${slug}/${MEMBERSHIPS_ENDPOINT}/`, data: { user: { user_id: userId } } }),
+  deleteMembership: (slug: string, userId: string) =>
+    IFetch<RequestResponse>({ method: 'DELETE', url: `${GROUPS_ENDPOINT}/${slug}/${MEMBERSHIPS_ENDPOINT}/${userId}/` }),
   updateMembership: (slug: string, userId: string, data: { membership_type: MembershipType }) =>
-    IFetch<Membership>({ method: 'PUT', url: `group/${slug}/membership/${userId}/`, data }),
+    IFetch<Membership>({ method: 'PUT', url: `${GROUPS_ENDPOINT}/${slug}/${MEMBERSHIPS_ENDPOINT}/${userId}/`, data }),
 
   //Group
-  getGroups: () => IFetch<Group[]>({ method: 'GET', url: `group/` }),
-  getGroup: (slug: string) => IFetch<Group>({ method: 'GET', url: `group/${slug}/` }),
-  updateGroup: (slug: string, data: Group) => IFetch<Group>({ method: 'PUT', url: `group/${slug}/`, data }),
+  getGroups: () => IFetch<Group[]>({ method: 'GET', url: `${GROUPS_ENDPOINT}/` }),
+  getGroup: (slug: string) => IFetch<Group>({ method: 'GET', url: `${GROUPS_ENDPOINT}/${slug}/` }),
+  updateGroup: (slug: string, data: Group) => IFetch<Group>({ method: 'PUT', url: `${GROUPS_ENDPOINT}/${slug}/`, data }),
 
   // Pages
-  getPageTree: () => IFetch<PageTree>({ method: 'GET', url: `page/tree/` }),
-  getPage: (path: string) => IFetch<Page>({ method: 'GET', url: `page/${path}` }),
-  getPages: (filters: any) => IFetch<PaginationResponse<PageChildren>>({ method: 'GET', url: `page/`, data: filters }),
-  createPage: (data: PageRequired) => IFetch<Page>({ method: 'POST', url: `page/`, data }),
-  updatePage: (path: string, data: Partial<Page>) => IFetch<Page>({ method: 'PUT', url: `page/${path}`, data }),
-  deletePage: (path: string) => IFetch<RequestResponse>({ method: 'DELETE', url: `page/${path}` }),
+  getPageTree: () => IFetch<PageTree>({ method: 'GET', url: `${PAGES_ENDPOINT}/tree/` }),
+  getPage: (path: string) => IFetch<Page>({ method: 'GET', url: `${PAGES_ENDPOINT}/${path}` }),
+  getPages: (filters: any) => IFetch<PaginationResponse<PageChildren>>({ method: 'GET', url: `${PAGES_ENDPOINT}/`, data: filters }),
+  createPage: (data: PageRequired) => IFetch<Page>({ method: 'POST', url: `${PAGES_ENDPOINT}/`, data }),
+  updatePage: (path: string, data: Partial<Page>) => IFetch<Page>({ method: 'PUT', url: `${PAGES_ENDPOINT}/${path}`, data }),
+  deletePage: (path: string) => IFetch<RequestResponse>({ method: 'DELETE', url: `${PAGES_ENDPOINT}/${path}` }),
 
   // File-upload
   uploadFile: (file: File | Blob) => IFetch<FileUploadResponse>({ method: 'POST', url: 'upload/', file }),