+++ src/containers/GroupOverview/components/GroupItem.tsx
@@ -1,29 +1,24 @@
 import { Link } from 'react-router-dom';
-import { Group } from 'types/Types';
+import { GroupList } from 'types/Types';
 import URLS from 'URLS';
 
 // Material UI
-import { ButtonBase, Typography } from '@material-ui/core';
-import { makeStyles } from '@material-ui/core/styles';
+import { makeStyles, ButtonBase, Typography } from '@material-ui/core';
+import Skeleton from '@material-ui/lab/Skeleton';
 
 // Icons
 import MembersIcon from '@material-ui/icons/PersonRounded';
 
 const useStyles = makeStyles((theme) => ({
   container: {
-    backgroundColor: theme.palette.background.default,
+    width: '100%',
+    backgroundColor: theme.palette.background.paper,
     border: `${theme.palette.borderWidth} solid ${theme.palette.divider}`,
     borderRadius: theme.shape.borderRadius,
     display: 'flex',
     flexDirection: 'column',
     height: '100%',
     minHeight: 90,
-    '&:hover > h3': {
-      transform: 'scale(1.1)',
-    },
-  },
-  group: {
-    transition: 'transform .2s',
   },
   leader: {
     display: 'flex',
@@ -41,17 +36,15 @@ const useStyles = makeStyles((theme) => ({
 }));
 
 export type GroupItemProps = {
-  group: Group;
+  group: GroupList;
 };
 
 const GroupItem = ({ group }: GroupItemProps) => {
   const classes = useStyles();
 
   return (
-    <ButtonBase className={classes.container} component={Link} to={`${URLS.groups}${group.slug}/`}>
-      <Typography className={classes.group} variant='h3'>
-        {group.name}
-      </Typography>
+    <ButtonBase className={classes.container} component={Link} focusRipple to={`${URLS.groups}${group.slug}/`}>
+      <Typography variant='h3'>{group.name}</Typography>
       {group.leader && (
         <div className={classes.leader}>
           <MembersIcon className={classes.icon} />
@@ -65,3 +58,16 @@ const GroupItem = ({ group }: GroupItemProps) => {
 };
 
 export default GroupItem;
+
+export const GroupItemLoading = () => {
+  const classes = useStyles();
+  return (
+    <ButtonBase className={classes.container} focusRipple>
+      <Skeleton width={100} />
+      <div className={classes.leader}>
+        <MembersIcon className={classes.icon} />
+        <Skeleton className={classes.name} width={120} />
+      </div>
+    </ButtonBase>
+  );
+};