+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 
 - ✨ **Filtrering**. Det er mulig å finne tidligere arrangementer og annonser ved å justere filtrering.
 - ✨ **Pages**. Lagt til mulighet til å søke innad i pages.
+- ✨ **Profil**. Pagination på arrangementer og badges, samt lagt til dine grupper.
 - 🎨 **Nyheter**. Nyheter viser nå hvilket år de ble publisert
 - 🦟 **Sidebar**. Lukker sidebar ved klikk til annen side innad i pages.